Annotation update

+++ annotations/nocedit.exe/pseudocode/src/core/setutil.cpp/C3DSLight_exportS3D_FUN_00587850.cpp
@@ -11,7 +11,8 @@ void __cdecl core_setutil_cpp_C3DSLight_exportS3D_FUN_00587850(C3DSLight *this_p
 {
   _fprintf(file_handle,"\"%s\",%d,%g,%g,%g,%g,%g,%g ",this_ptr->name,this_ptr->light_type,
              (double)(this_ptr->pos).x,(double)(this_ptr->pos).y,(double)(this_ptr->pos).z,
-             (double)(this_ptr->color).r,(double)(this_ptr->color).g,(double)(this_ptr->color).b);
+             (double)(float)(this_ptr->color).r,(double)(float)(this_ptr->color).g,
+             (double)(float)(this_ptr->color).b);
   if (this_ptr->light_type == 0) {
     _fprintf(file_handle,",%g,%g,%g\n",(double)(this_ptr->orient).x,
                (double)(this_ptr->orient).z,(double)(this_ptr->orient).y);
